Fix up the code to use ABSL.
Now it correctly sets things on my RDA5807. The kit
still doesn't work, which I can't fully figure out. But the
issue is not with the microcontroller at least!

I can hear interference from the Raspberry Pi I'm
running this from, so the radio is not totally busted.
But I don't hear any static from the antennae, so
there must still be something I'm missing.